+++ app/memcing.js
@@ -8,22 +8,18 @@
 /* jslint node: true */
 'use strict';
 
-var mFS       = require('fs'),
-    mReadline = require('readline'),
-    mHTTP     = require('http'),
-    mURL      = require('url'),
+var mReadline = require('readline'),
     mUtilex   = require('utilex'),
-    mQ        = require('q'),
     mCache    = require('./cache'),
     mHelp     = require('./help'),
     mConfig   = require('./config'),
-    mRegex    = require('./regex'),
-    mQS       = require('querystring')
+    mRest     = require('./rest')
 ;
 
 // Init vars
 var gConfig   = mConfig().get(),
-    gCache    = mCache(gConfig.cache)
+    gCache    = mCache(gConfig.cache),
+    gRest     = mRest(gConfig.listen, gCache)
 ;
 
 // Check whether help or not
@@ -34,7 +30,7 @@ if(gConfig.loadFile) {
   gCache.loadFile(gConfig.loadFile).then(function() { // load file
     if(gConfig.listen.http.isEnabled || gConfig.isIactive) {
       if(gConfig.listen.http.isEnabled) { // listen
-        cmdListen().then(function() {
+        gRest.listen().then(function() {
           if(gConfig.isIactive) cmdIactive(); // interactive mode
         }, function(err) {
           mUtilex.tidyLog(err);
@@ -52,7 +48,7 @@ if(gConfig.loadFile) {
   });
 } else if(gConfig.isIactive) {
   if(gConfig.listen.http.isEnabled) { // listen
-    cmdListen().then(function() {
+    gRest.listen().then(function() {
       cmdIactive(); // interactive mode
     }, function(err) {
       mUtilex.tidyLog(err);
@@ -168,152 +164,4 @@ function cmdIactive() {
       }
     });
   });
-}
-
-// Execute listen commands.
-function cmdListen() {
-
-  // Init vars
-  var deferred  = mQ.defer(),
-      hostname  = gConfig.listen.http.hostname || 'localhost',
-      port      = gConfig.listen.http.port || 12080,
-      resHdr    = {'Content-Type': 'application/json'}
-  ;
-
-  // Init http
-  var server = mHTTP.createServer(function(req, res) {
-
-    var up = mURL.parse(req.url, true, false);
-    //console.log(up); // for debug
-
-    if(up && up.pathname) {
-      var pathAry = up.pathname.split('/');
-      //console.log(pathAry); // for debug
-
-      if(pathAry[1] == 'entries') {
-
-        if(pathAry[2]) {
-          // element
-
-          var tElem = (mRegex.number.test(pathAry[2]) && !isNaN(pathAry[2]/1)) ? pathAry[2]/1 : pathAry[2];
-
-          if(req.method == 'GET') {
-            var cg = gCache.get(tElem);
-            if(cg) {
-              res.writeHead(200, resHdr);
-              res.end(JSON.stringify(cg));
-            } else {
-              res.writeHead(404, resHdr);
-              res.end(JSON.stringify({code: '404', message: 'Not Found'}));
-            }
-          } else if(req.method == 'PUT' || req.method == 'POST') {
-            var bodyAry = [],
-                dLen    = 0,
-                dLmt    = gCache.sizeOfPerEntry()*2
-            ;
-
-            req.on('data', function(chunk) {
-              //console.log(chunk.length + ' - ' + dLen + ' - ' + dLmt); // for debug
-              if(dLen >= dLmt) {
-                return false;
-              } else {
-                dLen += chunk.length;
-                bodyAry.push(chunk);
-              }
-            });
-
-            req.on('end', function() {
-              if(dLen >= dLmt) {
-                res.writeHead(413, resHdr);
-                res.end(JSON.stringify({code: '413', message: 'Request Entity Too Large'}));
-              } else {
-                if(req.headers['content-type'] == 'application/x-www-form-urlencoded') {
-                  var qsp     = mQS.parse(bodyAry.join()),
-                      eVal    = (qsp && qsp.val && mRegex.number.test(qsp.val) && !isNaN(qsp.val/1)) ? qsp.val/1 : ((qsp && qsp.val) ? qsp.val : null),
-                      eExp    = (qsp && qsp.exp) ? qsp.exp : null,
-                      setTrig = true
-                  ;
-
-                  if(req.method == 'POST') {
-                    var cg = gCache.get(tElem);
-                    if(cg) {
-                      res.writeHead(409, resHdr);
-                      res.end(JSON.stringify({code: '409', message: 'Conflict', entry: cg}));
-                      setTrig = false;
-                    }
-                  }
-
-                  if(setTrig === true) {
-                    var cs = gCache.set(tElem, eVal, eExp);
-                    if(!cs.error) {
-                      res.writeHead(200, resHdr);
-                      res.end(JSON.stringify(gCache.get(tElem)));
-                    } else {
-                      res.writeHead(400, resHdr);
-                      res.end(JSON.stringify({code: '400', message: cs.error}));
-                    }
-                  }
-                } else {
-                  res.writeHead(400, resHdr);
-                  res.end(JSON.stringify({code: '400', message: 'Bad Request (Use `application/x-www-form-urlencoded` for PUT/POST)'}));
-                }
-              }
-            });
-          } else if(req.method == 'DELETE') {
-            gCache.del(tElem);
-            res.writeHead(200, resHdr);
-            res.end();
-          } else {
-            res.writeHead(405, resHdr);
-            res.end(JSON.stringify({code: '405', message: 'Method Not Allowed'}));
-          }
-        } else {
-          // collection
-          
-          if(req.method == 'GET') {
-            res.writeHead(200, resHdr);
-
-            if(gCache.numOfEntry() > 0) {
-              var cd = gCache.entries(),
-                  cc = ''
-              ;
-              res.write('[');
-              for(var key in cd) {
-                res.write(cc + '\n' + JSON.stringify(cd[key]));
-                if(!cc) cc = ',';
-              }
-              res.write('\n]');
-            } else {
-              res.write('[]');
-            }
-
-            res.end();
-          } else if(req.method == 'DELETE') {
-            gCache.drop();
-            res.writeHead(200, resHdr);
-            res.end();
-          } else {
-            res.writeHead(405, resHdr);
-            res.end(JSON.stringify({code: '405', message: 'Method Not Allowed'}));
-          }
-        }
-      } else if(pathAry[1]) {
-        res.writeHead(501, resHdr);
-        res.end(JSON.stringify({code: '501', message: 'Not Implemented'}));
-      } else {
-        res.writeHead(200, resHdr);
-        res.end();
-      }
-    }
-  }).listen(port, hostname, function() {
-    mUtilex.tidyLog('Server is listening on ' + server.address().address + ':' + server.address().port);
-    deferred.resolve();
-  });
-
-  server.on('error', function (e) {
-    deferred.reject(e);
-  });
-
-
-  return deferred.promise;
 }